requirements.txt
streamlit
telethon
pandas
openpyxl
python-dotenv 